pipa: overlay: Mark that device is not voice capable
"Voice" capable means that the device supports phone calls over the telephony network. So for this device, this should be false.

The impetus for this is that with voice capability defaulting to true, the Multiuser pannel asks the owner whether secondary users should be allowed to make phone calls/sms, which is inapplicable. Changing the value to false should get rid of this option.

Bug: 248440908
Change-Id: If5249bc943c96f88c2ddd26880fe0626bb53592e
Test: manual
(cherry picked from commit 699275b04d85dde5645034cdeca19207b0bde632)
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>