pipa: overlay: Decrease the touch slop to prevent unintentional long presses
On this device, the default touch slop value of 8dp is too large. It corresponds to a physical distance of ~ 1.5mm.
This results in long presses getting triggered too easily.

Reduce the touch slop to 5 dp, which is about ~ 1mm, and closer to a reference device that doesn't exhibit the long-press issue.

Bug: 270835406
Change-Id: I1973e7831f527f2f654661bd4917f48992c74b61
Test: use a test app for long presses
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>